Moved and renamed loghttp v1 package
