+++ pkg/loghttp/labels.go
@@ -1,4 +1,4 @@
-package v1
+package loghttp
 
 import "github.com/prometheus/prometheus/promql"
