+++ pkg/loghttp/marshal_test.go
@@ -1,4 +1,4 @@
-package v1
+package loghttp
 
 import (
 	"bytes"
