+++ pkg/loghttp/tail.go
@@ -1,4 +1,4 @@
-package v1
+package loghttp
 
 import (
 	"encoding/json"
